move extractors into separate files
