+++ dropshot/tests/fail/bad_endpoint5.stderr
@@ -16,7 +16,7 @@ error[E0277]: the trait bound `for<'de> QueryParams: serde::de::Deserialize<'de>
            and $N others
    = note: required for `QueryParams` to implement `serde::de::DeserializeOwned`
 note: required by a bound in `dropshot::Query`
-  --> src/extractor/mod.rs
+  --> src/extractor/query.rs
    |
    | pub struct Query<QueryType: DeserializeOwned + JsonSchema + Send + Sync> {
    |                             ^^^^^^^^^^^^^^^^ required by this bound in `dropshot::Query`
